main.rs: add handling for paths with symbolic links
Root mountpoint directories that contain a symbolic link are not
handled correctly.

For example, /var/log/ -> /var/volatile/log, in this case when we
mount something to /var/log/... the kernel records the mountpoint as
/var/volatile/log/..., so when we try to unmount /var/log/... it fails.

This problem is fixed by fully expanding the mountpoint path (with
canonicalize) before checking and unmounting the mountpoint.

Change-type: patch
Changelog-entry: main.rs: add handling for paths with symbolic links
Signed-off-by: Mark Corbin <mark@balena.io>